third_party/blink/web_tests/external/wpt/html/syntax/charset/inheritance-bogus-meta-utf-8-expected.txt

This is a testharness.js-based test.
[FAIL] Child with bogus Content-Type charset
  assert_equals: expected "\\ufffd\\n" but got "¢\\n"
Harness: the test ran to completion.
